refactor of the `vkldap` module
This commit refactors the code present in the `vkldap` module since it
was already too much code in a single file.

It also does some naming changes to better match the code behavior.
