+++ src/commands.rs
@@ -4,7 +4,7 @@ use valkey_module::{
     Context, ValkeyError, ValkeyResult, ValkeyString, ValkeyValue, redisvalue::ValkeyValueKey,
 };
 
-use crate::vkldap::{VkLdapServerStatus, get_servers_health_status};
+use crate::vkldap::{get_servers_health_status, server::VkLdapServerStatus};
 
 pub fn ldap_status_command(_ctx: &Context, args: Vec<ValkeyString>) -> ValkeyResult {
     if args.len() > 1 {
@@ -18,14 +18,11 @@ pub fn ldap_status_command(_ctx: &Context, args: Vec<ValkeyString>) -> ValkeyRes
     let mut status_map: BTreeMap<ValkeyValueKey, ValkeyValue> = BTreeMap::new();
 
     for server in servers_health.iter() {
-        let status = match &server.status {
+        let status = match server.get_status() {
             VkLdapServerStatus::HEALTHY => "healthy".to_string(),
             VkLdapServerStatus::UNHEALTHY(err_msg) => format!("unhealthy: {err_msg}"),
         };
-        let hostname = match server.url.host() {
-            Some(host) => host.to_string(),
-            None => server.url.to_string(),
-        };
+        let hostname = server.get_host_string();
         status_map.insert(
             ValkeyValueKey::BulkValkeyString(ValkeyString::create(None, hostname)),
             ValkeyValue::BulkString(status.to_string()),
